Create cash.c
Program that prompts the user for the number of cents that a customer is owed and then prints the smallest number of coins with which that change can be made. When making change, odds are you want to minimize the number of coins you’re dispensing for each customer, lest you run out (or annoy the customer!). Fortunately, computer science has given cashiers everywhere ways to minimize numbers of coins due: greedy algorithms.